ci/requirements_ci.txt
azureml-core
openai
pandas
